change specs to support multiple databases
extracted database schema to dedicate file
